+++ src/time_unit/mod.rs
@@ -15,48 +15,13 @@ pub use self::seconds::Seconds;
 pub use self::years::Years;
 
 use crate::error::*;
-use crate::ordinal::{Ordinal, OrdinalSet};
+use crate::ordinal::{Ordinal, OrdinalSet, OrdinalIter, OrdinalRangeIter};
 use crate::specifier::{RootSpecifier, Specifier};
 use crate::field::{Field, FromField};
 
 use std::borrow::Cow;
-use std::collections::btree_set;
 use std::iter;
 use std::ops::RangeBounds;
-
-pub struct OrdinalIter<'a> {
-    set_iter: btree_set::Iter<'a, Ordinal>,
-}
-
-impl<'a> Iterator for OrdinalIter<'a> {
-    type Item = Ordinal;
-    fn next(&mut self) -> Option<Ordinal> {
-        self.set_iter.next().copied()
-    }
-}
-
-impl<'a> DoubleEndedIterator for OrdinalIter<'a> {
-    fn next_back(&mut self) -> Option<Self::Item> {
-        self.set_iter.next_back().copied()
-    }
-}
-
-pub struct OrdinalRangeIter<'a> {
-    range_iter: btree_set::Range<'a, Ordinal>,
-}
-
-impl<'a> Iterator for OrdinalRangeIter<'a> {
-    type Item = Ordinal;
-    fn next(&mut self) -> Option<Ordinal> {
-        self.range_iter.next().copied()
-    }
-}
-
-impl<'a> DoubleEndedIterator for OrdinalRangeIter<'a> {
-    fn next_back(&mut self) -> Option<Self::Item> {
-        self.range_iter.next_back().copied()
-    }
-}
 
 /// Methods exposing a schedule's configured ordinals for each individual unit of time.
 /// # Example
@@ -182,17 +147,13 @@ where
         self.ordinals().contains(&ordinal)
     }
     fn iter(&self) -> OrdinalIter {
-        OrdinalIter {
-            set_iter: TimeUnitField::ordinals(self).iter(),
-        }
+        OrdinalIter::new(TimeUnitField::ordinals(self))
     }
     fn range<R>(&'_ self, range: R) -> OrdinalRangeIter<'_>
     where
         R: RangeBounds<Ordinal>,
     {
-        OrdinalRangeIter {
-            range_iter: TimeUnitField::ordinals(self).range(range),
-        }
+        OrdinalRangeIter::new(TimeUnitField::ordinals(self), range)
     }
     fn count(&self) -> u32 {
         self.ordinals().len() as u32
